Allow advanced setup for stylelint core and plugin pack authors
